Restructured JSON converters, fixed long-to-string JSON.
