+++ UGemini/Packages/com.uralstech.ugemini/Runtime/Scripts/Data/Files/GeminiFileVideoMetaData.cs
@@ -1,6 +1,7 @@
 ﻿using Newtonsoft.Json;
 using Newtonsoft.Json.Serialization;
 using System;
+using Uralstech.UGemini.JsonConverters;
 
 namespace Uralstech.UGemini.FileAPI
 {
