+++ UGemini/Packages/com.uralstech.ugemini/Runtime/Scripts/Data/Models/Caching/Create/GeminiCachedContentCreationData.cs
@@ -2,6 +2,7 @@ using Newtonsoft.Json;
 using Newtonsoft.Json.Serialization;
 using System;
 using System.ComponentModel;
+using Uralstech.UGemini.JsonConverters;
 using Uralstech.UGemini.Models.Content;
 using Uralstech.UGemini.Models.Generation.Tools.Declaration;
