+++ UGemini/Packages/com.uralstech.ugemini/Runtime/Scripts/Data/Models/Tuning/Patch/GeminiTunedModelPatchData.cs
@@ -1,6 +1,7 @@
 using Newtonsoft.Json;
 using Newtonsoft.Json.Serialization;
 using System.ComponentModel;
+using Uralstech.UGemini.JsonConverters;
 
 namespace Uralstech.UGemini.Models.Tuning
 {
