+++ UGemini/Packages/com.uralstech.ugemini/Runtime/Scripts/JsonConverters.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 15169ad2329617544b377b67f6880b70
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
