+++ UGemini/Packages/com.uralstech.ugemini/Runtime/Scripts/JsonConverters/GeminiLongArrayToStringArrayJsonConverter.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 81e346c66d480604d8a7efbe33b64340
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
