+++ UGemini/Packages/com.uralstech.ugemini/Runtime/Scripts/JsonConverters/GeminiLongToStringJsonConverter.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 43f812835e769204cae2631e63f320d5
+guid: dfd2fe2f510452547b8e2b1558c2839b
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
