+++ UGemini/Packages/com.uralstech.ugemini/Runtime/Scripts/JsonConverters/GeminiModelIdStringConverter.cs.meta
