+++ UGemini/Packages/com.uralstech.ugemini/Runtime/Scripts/JsonConverters/GeminiSecondsToTimeSpanJsonConverter.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f097c2b4d45ac1343b882cd4967bce65
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
